chore: Make psalm run with lowest supported PHP version
To have consistent psalm results the target PHP version should be set
instead of guessed from currently running version.
Currently Nextcloud 28 is supported so lowest PHP version is 8.0.

Also the schemaLocation was set to a non-existing path, as the phar bundle
is used the only available location would be the online version.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>